dodavanje/updateanje property i client entitya u store nakon uspješnog dodavanja u bazu

+++ src/models/Clients.ts
@@ -2,6 +2,8 @@ import type { UserData } from "../types";
 
 import { pb } from "../PocketBaseInit";
 import { ClientResponseError } from "pocketbase";
+import { handlePocketbaseError } from "./errorHandling";
+import { addClientToStore, updateClientInStore } from "../stores/actions";
 
 // EXAMPLE DATA
 // {
@@ -37,7 +39,7 @@ export async function getUsers(): Promise<UserData[]> {
     return data;
   } catch (err) {
     console.error("Error fetching users:", err);
-    throw handleClientError(err);
+    throw handlePocketbaseError(err);
   }
 }
 
@@ -46,59 +48,34 @@ export async function createUser(data: Partial<UserData>): Promise<void> {
   try {
     const collection = pb.collection<UserData>("Clients");
     if (data.id) {
+      // Update existing user
       const res = await collection.update(data.id, data);
       console.log("User updated:", res);
+      updateClientInStore(res);
     } else {
+      // Create new user
       const res = await collection.create(data);
       console.log("User created:", res);
+      addClientToStore(res);
     }
   } catch (err) {
     console.error("Error adding/updating user:", err);
-    throw handleClientError(err);
+    throw handlePocketbaseError(err);
   }
 }
 
 // DELETE
 export async function deleteUser(id: string): Promise<void> {
+  if (!id) {
+    throw new Error("Id is required");
+  }
+
   try {
     const collection = pb.collection<UserData>("Clients");
     const res = await collection.delete(id);
     console.log("User deleted:", res);
   } catch (err) {
     console.error("Error deleting user:", err);
-    throw handleClientError(err);
+    throw handlePocketbaseError(err);
   }
-}
-
-// ERROR HANDLING
-function handleClientError(err: unknown): Error | Record<string, string> {
-  if (err instanceof ClientResponseError) {
-    console.log("ClientResponseError details:");
-    if (err.response?.data && typeof err.response.data === "object") {
-      return parseErrorResponse(err.response.data);
-    } else if (err.message) {
-      return { error: err.message };
-    }
-  }
-  console.log(err);
-  return err as Error;
-}
-
-function parseErrorResponse(data: Record<string, unknown>): Record<string, string> {
-  const errorObject: Record<string, string> = {};
-  Object.entries(data).forEach(([key, value]) => {
-    if (isErrorObject(value)) {
-      const { code = "Unknown code", message = "Unknown message" } = value;
-      errorObject[key] = `${code} - ${message}`;
-    } else {
-      errorObject[key] = String(value);
-    }
-  });
-  return errorObject;
-}
-
-function isErrorObject(value: unknown): value is { code?: string; message?: string } {
-  return (
-    typeof value === "object" && value !== null && ("code" in value || "message" in value)
-  );
 }
